Add files via upload
Onboard carbon capture and storage (CCS) is one potential way forward for reducing CO2 emission from shipping. Carbon is extracted from the ship´s exhaust gases, and stored onboard until these tanks can be emptied at a CCS reception facility, and further transported to the long term storage site.
The CCS tanks will require space onboard, and ultimately reduce the cargo carrier capacity of the vessel. This call for frequent unloading of CO2, and many facilities. On the other hand, such facilities are costly to establish and operate, calling for few facilities. Thus, we need to strike a balance.
In the basis model, we capture a vessel sailing along a fixed route with varying inter-port distances. The main decision variable is to determine the optimal location of reception facility points.